requirements.txt
attrs==26.1.0
cdsapi==0.7.7
certifi==2026.4.22
cftime==1.6.5
charset-normalizer==3.4.7
cloudpickle==3.1.2
colorama==0.4.6
ecmwf-datastores-client==0.5.1
filelock==3.29.0
fsspec==2026.3.0
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.47.0
MarkupSafe==3.0.3
mpmath==1.3.0
multiurl==0.3.7
netCDF4==1.7.4
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
xarray==2026.4.0
